Only add extension key in dev mode
